finish:DictData,DictType

+++ blog-service/src/main/java/top/ooovo/blog/service/modules/system/dal/mysql/permission/SysMenuMapper.java
@@ -22,7 +22,7 @@ public interface SysMenuMapper extends BaseMapperX<SysMenuDO> {
                 .eq("name", name));
     }
 
-    default Integer selectCountByParentId(Long parentId) {
+    default Long selectCountByParentId(Long parentId) {
         return selectCount(new QueryWrapperX<SysMenuDO>().eq("parent_id", parentId));
     }
